# vec to smallvec