Convert `Vec<Gc<Value>>` into `SmallVec<[Gc<Value>; 1]>`.

This does not add any dependencies, as it is already used by crossterm.

- **fixed merge conflict**
- **cargo fmt**