@asdfish 2025-02-24
I currently cannot run the benchmarks without running out of ram, but when I tested it earlier I got ~26s when I usually get ~36s before this change.