@asdfish 2025-02-24
# Before
```txt
     Running benches/fib.rs (target/release/deps/fib-da8f3a1f9e3edc32)
Gnuplot not found, using plotters backend
Benchmarking fib 10000: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 31.8s, or reduce sample count to 10.
fib 10000               time:   [324.28 ms 327.34 ms 330.39 ms]
                        change: [+0.9399% +2.3555% +3.9106%] (p = 0.00 < 0.05)
                        Change within noise threshold.
```
# After
```txt
     Running benches/fib.rs (target/release/deps/fib-dda3328bc6a14c28)
Gnuplot not found, using plotters backend
Benchmarking fib 10000: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 27.2s, or reduce sample count to 10.
fib 10000               time:   [271.77 ms 273.83 ms 275.86 ms]
                        change: [-0.5962% +0.4716% +1.5044%] (p = 0.39 > 0.05)
                        No change in performance detected.
```