@maplant 2025-04-23
Closing this for now, the source code has diverged so far and there are much stronger optimizations to follow